Further streamline job tab
